requirements.txt
torch>=2.5
torchvision>=0.20
numpy>=1.26
Pillow>=10.0
opencv-python>=4.9
tqdm>=4.66
albumentations>=1.4
segmentation-models-pytorch>=0.3.4
label-studio>=1.23
